Remove unnecessary elements of Service protocol
